+++ OnlineVideoLinks/HelpForm.cs
@@ -3,18 +3,25 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
 using System.Drawing;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace ExtraGameVideos
+namespace OnlineVideoLinks
 {
     public partial class HelpForm : Form
     {
         public HelpForm()
         {
             InitializeComponent();
+        }
+
+        private void HelpForm_Load(object sender, EventArgs e)
+        {
+            richTextBox1.BackColor = this.BackColor;
+            richTextBox1.Rtf = Properties.Resources.HelpDocument;
         }
     }
 }
